* renamed premake4_config/ to premake4_tools/
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/experimental/v2_99continuing/bzflag@21837 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ premake4_tools/utils.lua
@@ -152,6 +152,10 @@ function os.testcode(t)
 
   local result = os.execute(cmd)
 
+  if (result == 2) then
+    os.exit(2) -- SIGINT, bail
+  end
+
   os.remove(tmpOut)
   os.remove(tmpCpp)
